requirements.txt
requests>=2.28.0
pandas>=1.5.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
# 可选依赖（备选数据源）
# akshare>=1.10.0
# playwright>=1.40.0
